Port lite licence usage reporting task to celery
